Fix merge conflicts

+++ goobi-viewer-core/src/main/java/io/goobi/viewer/managedbeans/AdminBean.java
@@ -235,7 +235,9 @@ public class AdminBean implements Serializable {
      * @throws io.goobi.viewer.exceptions.DAOException if any.
      */
     public List<User> getAllUsers() throws DAOException {
-        return DataManager.getInstance().getDao().getAllUsers(true);
+        List<User> users = DataManager.getInstance().getDao().getAllUsers(true);
+        Collections.sort(users);
+        return users;
     }
 
     /**
@@ -664,10 +666,12 @@ public class AdminBean implements Serializable {
                         break;
                     case "delete":
                         logger.trace("Deleting UserRole: {}", userRole);
-                        if (DataManager.getInstance().getDao().deleteUserRole(userRole)) {
-                            Messages.info("deletedSuccessfully");
-                        } else {
-                            Messages.error("deleteFailure");
+                        if(userRole.getId() != null) {                            
+                            if (DataManager.getInstance().getDao().deleteUserRole(userRole)) {
+                                Messages.info("deletedSuccessfully");
+                            } else {
+                                Messages.error("deleteFailure");
+                            }
                         }
                         break;
                     default:
